Work on simplifying the dynamic string API.
Functions no longer return errno_t. Instead, the state of the string is
tracked within the string's metadata alongside the length and capacity
variables. The return value is now a c89str. When an error occurs, such
as ENOMEM, a pointer to the original input string is returned. This
system allows the caller to easily keep track of the original string in
case they need to do cleanup.

When a dynamic string is in an error state, any operations will abort
without modifying the string. It's up to the caller to inspect the
error status of the string with `c89str_result()`.